+++ libraries/base/src/GHC/Conc/IO.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE CPP #-}
 {-# OPTIONS_HADDOCK not-home #-}
 
 -- |
@@ -7,7 +8,7 @@
 -- License     :  see libraries/base/LICENSE
 --
 -- Maintainer  :  ghc-devs@haskell.org
--- Stability   :  internal
+-- Stability   :  deprecated (<https://github.com/haskell/core-libraries-committee/issues/393>)
 -- Portability :  non-portable (GHC extensions)
 --
 -- Basic concurrency stuff.
@@ -18,7 +19,12 @@
 -- change rapidly without much warning.
 --
 
+#if __GLASGOW_HASKELL__ >= 1002
+#error "GHC.Conc.IO should be removed in GHC 10.02."
+#endif
+
 module GHC.Conc.IO
+  {-# DEPRECATED "GHC.Conc.IO is deprecated and will be removed in GHC 10.02. Please use GHC.Conc where possible and otherwise ask for the inclusion of needed functionality into ghc-experimental via the GHC issue tracker." #-}
     (ensureIOManagerIsRunning,
      ioManagerCapabilitiesChanged,
      interruptIOManager,