+++ libraries/base/src/GHC/Exception/Type.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE CPP #-}
 {-# LANGUAGE Safe #-}
 {-# OPTIONS_HADDOCK not-home #-}
 
@@ -7,13 +8,18 @@
 -- License     :  see libraries/base/LICENSE
 --
 -- Maintainer  :  cvs-ghc@haskell.org
--- Stability   :  internal
+-- Stability   :  deprecated (<https://github.com/haskell/core-libraries-committee/issues/393>)
 -- Portability :  non-portable (GHC extensions)
 --
 -- Exceptions and exception-handling functions.
 --
 
+#if __GLASGOW_HASKELL__ >= 1002
+#error "GHC.Exception.Type should be removed in GHC 10.02."
+#endif
+
 module GHC.Exception.Type
+  {-# DEPRECATED "GHC.Exception.Type is deprecated and will be removed in GHC 10.02. Please use Control.Exception where possible and otherwise ask for the inclusion of needed functionality into ghc-experimental via the GHC issue tracker." #-}
        ( Exception(..)    -- Class
        , SomeException(..), ArithException(..)
        , divZeroException, overflowException, ratioZeroDenomException