+++ libraries/base/src/GHC/IO/Device.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE CPP #-}
 {-# LANGUAGE Safe #-}
 
 -- |
@@ -6,7 +7,7 @@
 -- License     :  see libraries/base/LICENSE
 --
 -- Maintainer  :  libraries@haskell.org
--- Stability   :  internal
+-- Stability   :  deprecated (<https://github.com/haskell/core-libraries-committee/issues/393>)
 -- Portability :  non-portable
 --
 -- Type classes for I/O providers.
@@ -17,7 +18,12 @@
 -- change rapidly without much warning.
 --
 
+#if __GLASGOW_HASKELL__ >= 1002
+#error "GHC.IO.Device should be removed in GHC 10.02."
+#endif
+
 module GHC.IO.Device (
+  {-# DEPRECATED "GHC.IO.Device is deprecated and will be removed in GHC 10.02. Please ask for the inclusion of needed functionality into ghc-experimental via the GHC issue tracker." #-}
         RawIO(..),
         IODevice(..),
         IODeviceType(..),