+++ libraries/base/src/GHC/InfoProv.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE CPP #-}
 {-# LANGUAGE Safe #-}
 
 -- |
@@ -6,7 +7,7 @@
 -- License     :  see libraries/base/LICENSE
 --
 -- Maintainer  :  ghc-devs@haskell.org
--- Stability   :  internal
+-- Stability   :  deprecated (<https://github.com/haskell/core-libraries-committee/issues/393>)
 -- Portability :  non-portable (GHC Extensions)
 --
 -- Access to GHC's info-table provenance metadata.
@@ -19,7 +20,12 @@
 -- @since 4.18.0.0
 --
 
+#if __GLASGOW_HASKELL__ >= 1002
+#error "GHC.InfoProv should be removed in GHC 10.02."
+#endif
+
 module GHC.InfoProv
+  {-# DEPRECATED "GHC.InfoProv is deprecated and will be removed in GHC 10.02. Please ask for the inclusion of needed functionality into ghc-experimental via the GHC issue tracker." #-}
     ( InfoProv(..)
     , ipLoc
     , ipeProv