+++ libraries/base/src/GHC/TopHandler.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE CPP #-}
 {-# LANGUAGE Safe #-}
 {-# OPTIONS_HADDOCK not-home #-}
 
@@ -8,7 +9,7 @@
 -- License     :  see libraries/base/LICENSE
 --
 -- Maintainer  :  ghc-devs@haskell.org
--- Stability   :  internal
+-- Stability   :  deprecated (<https://github.com/haskell/core-libraries-committee/issues/393>)
 -- Portability :  non-portable (GHC Extensions)
 --
 -- Support for catching exceptions raised during top-level computations
@@ -20,7 +21,12 @@
 -- change rapidly without much warning.
 --
 
+#if __GLASGOW_HASKELL__ >= 1002
+#error "GHC.TopHandler should be removed in GHC 10.02."
+#endif
+
 module GHC.TopHandler
+  {-# DEPRECATED "GHC.TopHandler is deprecated and will be removed in GHC 10.02. Please use GHC.Conc where possible and otherwise ask for the inclusion of needed functionality into ghc-experimental via the GHC issue tracker." #-}
     (runMainIO,
      runIO,
      runIOFastExit,